To serve people with no external SD card, we now also check for the .magisk trigger file in the standard internal Download directory (in addition to the root of the external SD card). We also still check for the string _magisk as part of the APGK zip file name.
If triggered, we now also look in the standard internal Download
directory (in addition to a ./Magisk directory on the external SD card)
for a suitable Magisk zip file to install.

To make the system more foolproof, after looking for versioned files
matching the pattern Magisk-v* (in which case we use the latest
according to lexical sort order), we then also look for files called
Magisk.zip and magisk.zip in the aforementioned locations. If we find
any, we use the one with the most recent timestamp.

If still no files are found, chaining of Magisk is abandoned.